test: 기존 동시성 테스트 클래스 분리
